refactor(api): move yesman.epicfight.api.client.online.* -> yesman.epicfight.client.online.*
