+++ src/main/java/yesman/epicfight/client/online/EpicFightServerConnectionHelper.java
@@ -1,19 +1,4 @@
-package yesman.epicfight.api.client.online;
-
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.http.HttpClient;
-import java.net.http.HttpResponse;
-import java.security.KeyManagementException;
-import java.security.NoSuchAlgorithmException;
-import java.text.MessageFormat;
-import java.time.Duration;
-import java.util.function.BiConsumer;
-
-import javax.net.ssl.SSLContext;
+package yesman.epicfight.client.online;
 
 import net.minecraft.Util;
 import net.minecraftforge.api.distmarker.Dist;
@@ -21,6 +6,16 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 import yesman.epicfight.api.client.model.Mesh;
 import yesman.epicfight.api.utils.ParseUtil;
 import yesman.epicfight.main.EpicFightMod;
+
+import javax.net.ssl.SSLContext;
+import java.io.*;
+import java.net.http.HttpClient;
+import java.net.http.HttpResponse;
+import java.security.KeyManagementException;
+import java.security.NoSuchAlgorithmException;
+import java.text.MessageFormat;
+import java.time.Duration;
+import java.util.function.BiConsumer;
 
 @OnlyIn(Dist.CLIENT)
 public class EpicFightServerConnectionHelper {
