+++ src/main/java/yesman/epicfight/client/online/EpicSkins.java
@@ -1,4 +1,4 @@
-package yesman.epicfight.api.client.online;
+package yesman.epicfight.client.online;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStreamReader;
